Constructed a desirable model with desired results
I have abandond the idea of searching for a best solution for the 50-50 ratio, it is clear that the following ratio proves to be completely useless with the present models. Instead, I have accpted 90-10 ratio and used the RandomForestClassifier to construct the model. The result came out with with 72% on the minority class. It isn't perfect, but this could be a start.